move the timeline components to a common place to be used by devices as well
