+++ frontend/src/pages/instance/VersionHistory/Timeline/index.vue
@@ -55,12 +55,12 @@ import FeatureUnavailable from '../../../../components/banners/FeatureUnavailabl
 import FeatureUnavailableToTeam from '../../../../components/banners/FeatureUnavailableToTeam.vue'
 import AssetDetailDialog from '../../../../components/dialogs/AssetDetailDialog.vue'
 import SnapshotEditDialog from '../../../../components/dialogs/SnapshotEditDialog.vue'
+import TimelineEvent from '../../../../components/version-history/timeline/TimelineEvent.vue'
 import { scrollTo } from '../../../../composables/Ux.js'
 import featuresMixin from '../../../../mixins/Features.js'
 import snapshotsMixin from '../../../../mixins/Snapshots.js'
 import SnapshotExportDialog from '../../../application/Snapshots/components/dialogs/SnapshotExportDialog.vue'
 
-import TimelineEvent from './components/TimelineEvent.vue'
 export default {
     name: 'HistoryTimeline',
     components: {
